spack.util.elide_list -> llnl.util.lang.elide_list
