+++ lib/spack/spack/stage.py
@@ -18,6 +18,7 @@ from typing import Dict  # novm
 
 from six import iteritems, string_types
 
+import llnl.util.lang
 import llnl.util.tty as tty
 from llnl.util.filesystem import (
     can_access,
@@ -34,7 +35,6 @@ import spack.error
 import spack.fetch_strategy as fs
 import spack.mirror
 import spack.paths
-import spack.util
 import spack.util.lock
 import spack.util.path as sup
 import spack.util.pattern as pattern
@@ -857,7 +857,7 @@ def get_checksums_for_versions(
     tty.msg('Found {0} version{1} of {2}:'.format(
             num_ver, '' if num_ver == 1 else 's', name),
             '',
-            *spack.util.elide_list(
+            *llnl.util.lang.elide_list(
                 ['{0:{1}}  {2}'.format(str(v), max_len, url_dict[v])
                  for v in sorted_versions]))
     print()
